requirements.txt
djangorestframework
psycopg2-binary
Pillow
python-decouple